requirements.txt
keras
numpy
pandas
scikit-learn
torch
torchvision
matplotlib
seaborn
jupyter
joblib
torchvision
pillow
tqdm
transformers
tf-keras